Add files via upload
Here I am creating a web application that is based off the Productive app. The user will be able to add tasks, change their color and add a specific logo next to them. Right now I have created the basic design and a popup window for adding tasks to the list.